fix(library): split plain functions out of utils.tsx to fix Fast Refresh
